+++ testdata/de/dm/intellij/liferay/language/xml/LiferayXPathInjectorTest/com/liferay/portal/kernel/xml/Document.java
@@ -1,0 +1,7 @@
+package com.liferay.portal.kernel.xml;
+
+public interface Document {
+
+	public String valueOf(String xPathExpression);
+
+}