+++ testdata/de/dm/intellij/liferay/language/xml/LiferayXPathInjectorTest/com/liferay/portal/kernel/xml/Node.java
@@ -1,0 +1,10 @@
+package com.liferay.portal.kernel.xml;
+
+import java.io.Serializable;
+import java.util.List;
+
+public interface Node extends Serializable {
+
+	public List<Node> selectNodes(String xPathExpression);
+
+}