+++ testdata/de/dm/intellij/liferay/language/xml/LiferayXPathInjectorTest/com/liferay/portal/kernel/xml/SAXReaderUtil.java
@@ -1,0 +1,9 @@
+package com.liferay.portal.kernel.xml;
+
+import com.liferay.portal.kernel.xml.Document;
+
+public interface SAXReaderUtil {
+
+	public Document read(String xml);
+
+}