+++ testdata/de/dm/intellij/liferay/language/xml/LiferayXPathInjectorTest/de/dm/helper/MyHelper.java
@@ -1,0 +1,12 @@
+package de.dm.helper;
+
+import com.liferay.portal.kernel.xml.Node;
+
+public class MyHelper {
+
+	public static void main(String[] args) {
+		Node node = null;
+
+		System.out.println(node.selectNodes("<caret>"));
+	}
+}