+++ testdata/de/dm/intellij/liferay/language/xml/LiferayXPathInjectorTest/sample.groovy
@@ -1,0 +1,5 @@
+import com.liferay.portal.kernel.xml.SAXReaderUtil
+
+def docXml = SAXReaderUtil.read("<root/>")
+
+def text = docXml.valueOf("t<caret>")